# backend/openapi.yml
openapi: 3.0.3
info:
  title: CrowdPulse - OpenAPI 3.0
  description: "CrowdPulse API specifics"
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
paths:
...
  /v1/tweets/list:
    get:
      tags: [v1]
      summary: Get the data of tweets.
      parameters:
        - name: dbs
          in: query
          description: MongoDB database names
          required: true
          schema:
            type: array
        - name: algorithm
          in: query
          description: all, sent-it, feel-it or hate-speech
          required: false
          schema:
            type: string
        - name: sentiment
          in: query
          description: all, positive, neutral or negative
          required: false
          schema:
            type: string
        - name: emotion
          in: query
          description: (only if algorithm = 'feel-it'), all, joy, sadness anger or fear
          required: false
          schema:
            type: string
        - name: hateSpeech
          in: query
          description: (only if algorithm = 'hate-speech'), all, acceptable, inappropriate, offensive or violent
          required: false
          schema:
            type: string
        - name: dateFrom
          in: query
          description: ISO date like 2022-01-09T00:00:00.000Z (dateTo required)
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          description: ISO date like 2022-01-09T00:00:00.000Z (dateFrom required)
          required: false
          schema:
            type: string
        - name: tags
          in: query
          description: A list of tags
          required: false
          schema:
            type: array
        - name: processedText
          in: query
          description: A list processed words
          required: false
          schema:
            type: array
        - name: hashtags
          in: query
          description: A list of hashtags (# not required)
          required: false
          schema:
            type: array
        - name: usernames
          in: query
          description: A list of usernames
          required: false
          schema:
            type: array
        - name: pageSize
          in: query
          description: The size of any page
          required: false
          schema:
            type: number
            default: 10
        - name: page
          in: query
          description: A number that represents the page
          required: false
          schema:
            type: number
            default: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1_tweets_list'
        '400':
          description: missing query parameter
        '500':
          description: Executing error
  # [v1] GET - /v1/map
...
components:
  schemas:
...
    v1_tweets_list:
      type: object
      example: {
        total: number,
        values: [
          {
            "author": string,
            "text": string,
            "lang": string,
            "sensitive": string,
            "created_at": date,
            "tags": Array
          }
        ]
      }
...